requirements.txt
chromadb==0.5.5
faiss_cpu==1.8.0.post1
Jinja2==3.1.4
langchain==0.2.13
langchain_community==0.2.12
langchain_openai==0.1.21
langchain_text_splitters==0.2.2
lingua==4.15.0
promptflow==1.12.0
promptflow_tools==1.4.0
streamlit==1.36.0
tiktoken==0.7.0
